refactor - LadderGame Ladder 와 합칩니당.
