+++ src/main/java/nextstep/ladder/domain/Ladder.java
@@ -1,5 +1,7 @@
 package nextstep.ladder.domain;
 
+import nextstep.ladder.ErrorMessage;
+
 import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -9,15 +11,30 @@ import static java.util.stream.Collectors.collectingAndThen;
 
 public class Ladder {
 
+    private static final String ALL_USER_KEY = "all";
     private LadderGoals ladderGoals;
     private List<LadderRow> ladderRows;
 
     public Ladder(Users users, LadderHeight height, LadderGoals ladderGoals, DirectionStrategy directionStrategy) {
+        throwIfNull(users, ladderGoals);
+        throwIfInvalidLadderGame(users, ladderGoals);
         this.ladderGoals = ladderGoals;
         this.ladderRows = IntStream.range(0, height.getValue())
                 .boxed()
                 .map(index -> new LadderRow(users, directionStrategy))
                 .collect(collectingAndThen(Collectors.toList(), Collections::unmodifiableList));
+    }
+
+    private void throwIfNull(Users users, LadderGoals ladderGoals) {
+        if (users == null || ladderGoals == null) {
+            throw new IllegalArgumentException(ErrorMessage.INVALID_LADDER_GAME);
+        }
+    }
+
+    private void throwIfInvalidLadderGame(Users users, LadderGoals ladderGoals) {
+        if (users.size() != ladderGoals.size()) {
+            throw new IllegalArgumentException(ErrorMessage.INVALID_LADDER_GAME);
+        }
     }
 
     public List<LadderRow> getLadderRows() {
@@ -35,4 +52,21 @@ public class Ladder {
     public LadderGoals getLadderGoals() {
         return ladderGoals;
     }
+
+    public List<LadderResult> start(String key, Users users) {
+        if (ALL_USER_KEY.equals(key)) {
+            return users.export()
+                    .stream()
+                    .map(item -> moveUser(item, users))
+                    .collect(Collectors.toList());
+        }
+        return Collections.singletonList(moveUser(new User(key), users));
+    }
+
+    private LadderResult moveUser(User user, Users users) {
+        int userIndex = users.indexOf(user);
+        LadderGoal ladderGoal = move(userIndex);
+        return new LadderResult(user, ladderGoal);
+    }
 }
+
