--- src/main/java/nextstep/ladder/domain/LadderGame.java
@@ -1,62 +1,0 @@
-package nextstep.ladder.domain;
-
-import nextstep.ladder.ErrorMessage;
-
-import java.util.Collections;
-import java.util.List;
-import java.util.stream.Collectors;
-
-public class LadderGame {
-
-    private static final String ALL_USER_KEY = "all";
-    private final Users users;
-    private final Ladder ladder;
-
-    public LadderGoals getLadderGoalBoard() {
-        return ladder.getLadderGoals();
-    }
-
-    public LadderGame(Users users, LadderHeight height, LadderGoals ladderGoals, DirectionStrategy directionStrategy) {
-        throwIfNull(users, ladderGoals);
-        throwIfInvalidLadderGame(users, ladderGoals);
-        this.users = users;
-        this.ladder = new Ladder(users, height, ladderGoals, directionStrategy);
-
-    }
-
-    private void throwIfNull(Users users, LadderGoals ladderGoals) {
-        if (users == null || ladderGoals == null) {
-            throw new IllegalArgumentException(ErrorMessage.INVALID_LADDER_GAME);
-        }
-    }
-
-    private void throwIfInvalidLadderGame(Users users, LadderGoals ladderGoals) {
-        if (users.size() != ladderGoals.size()) {
-            throw new IllegalArgumentException(ErrorMessage.INVALID_LADDER_GAME);
-        }
-    }
-
-    public List<LadderResult> start(String key) {
-        if (ALL_USER_KEY.equals(key)) {
-            return users.export()
-                    .stream()
-                    .map(this::moveUser)
-                    .collect(Collectors.toList());
-        }
-        return Collections.singletonList(moveUser(new User(key)));
-    }
-
-    private LadderResult moveUser(User user) {
-        int userIndex = users.indexOf(user);
-        LadderGoal ladderGoal = ladder.move(userIndex);
-        return new LadderResult(user, ladderGoal);
-    }
-
-    public Users getUsers() {
-        return users;
-    }
-
-    public Ladder getLadder() {
-        return ladder;
-    }
-}
